+++ app/upgrades/v16/upgrades.go
@@ -7,6 +7,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/types/module"
 	authkeeper "github.com/cosmos/cosmos-sdk/x/auth/keeper"
 	bankkeeper "github.com/cosmos/cosmos-sdk/x/bank/keeper"
+	govkeeper "github.com/cosmos/cosmos-sdk/x/gov/keeper"
 	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
 	"github.com/evmos/evmos/v16/precompiles/bech32"
 	osmosisoutpost "github.com/evmos/evmos/v16/precompiles/outposts/osmosis"
@@ -24,7 +25,8 @@ func CreateUpgradeHandler(
 	ek *evmkeeper.Keeper,
 	_ bankkeeper.Keeper,
 	inflationKeeper inflationkeeper.Keeper,
-	_ authkeeper.AccountKeeper,
+	ak authkeeper.AccountKeeper,
+	gk govkeeper.Keeper,
 ) upgradetypes.UpgradeHandler {
 	return func(ctx sdk.Context, _ upgradetypes.Plan, vm module.VersionMap) (module.VersionMap, error) {
 		logger := ctx.Logger().With("upgrade", UpgradeName)
@@ -45,12 +47,13 @@ func CreateUpgradeHandler(
 			logger.Error("failed to enable outposts", "error", err.Error())
 		}
 
-		// TODO: uncomment when ready
 		// Migrate the FeeCollector module account to include the Burner permission.
-		// if err := MigrateFeeCollector(ak, ctx); err != nil {
-		//	logger.Error("failed to migrate the fee collector", "error", err.Error())
-		// }
-		//
+		// This is required when including the postHandler to burn Cosmos Tx fees
+		if err := MigrateFeeCollector(ak, ctx); err != nil {
+			logger.Error("failed to migrate the fee collector", "error", err.Error())
+		}
+
+		// TODO: uncomment when ready
 		// if err := BurnUsageIncentivesPool(ctx, bankKeeper); err != nil {
 		//	logger.Error("failed to burn inflation pool", "error", err.Error())
 		// }
@@ -58,6 +61,10 @@ func CreateUpgradeHandler(
 		if err := UpdateInflationParams(ctx, inflationKeeper); err != nil {
 			logger.Error("failed to update inflation params", "error", err.Error())
 		}
+
+		// Remove the deprecated governance proposals from store
+		logger.Debug("deleting deprecated incentives module proposals...")
+		DeleteIncentivesProposals(ctx, gk, logger)
 
 		// recovery module is deprecated
 		logger.Debug("deleting recovery module from version map...")